[vcpkg] Fix issue #9916; `vcpkg upgrade` did not `load_tag_vars()`
